Timer: Update UI more frequently for reliability.
After the timer has run for 15+ minutes, the updates can sometimes start to seem irratic. Sometimes a value will jump from `21:01` to `21:03`, without showing `21:02`.

This is probably because the elapsed time has slightly drifted, and the 1 second window is no longer perfectly lined up with the timer interval. Updating more often should reduce that below the threshold where it is noticable. If this still happens, then reducing to `250` ms should help.